moved helper functions to Utils/index.js
